+++ CHANGELOG.md
@@ -4,11 +4,27 @@ All notable changes to the "zephyr-workbench" extension will be documented in th
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-<<<<<<< HEAD
-## [1.3.38]
+## [2.1.0]
 
-- Fix STM32CubeProgrammer install
-=======
+- Fix Debug Manager default reset and avoid scanning until the view is opened
+- Pre-fill “Select the application to debug” when launching Debug from an application
+- Add remove actions in Multi-build for build configs and default runner
+- Disable “Sample project” selection while samples are loading
+- Refresh Host Tools versions after project creation (no VS Code reload needed)
+- Refresh runner “path” from YAML immediately after install
+- Warn when initializing into a path that already have a west workspace
+- Ensure west manifest revision/version updates correctly after `west update`
+- Add Minimal/Full option when creating west workspace from template
+
+## [2.0.4]
+
+- Fetch toolchains per version remotely with inline spinner feedback
+
+## [2.0.3]
+
+- Update documentation to standardize the new extension name Workbench for Zephyr
+- Adjust website and documentation titles for consistency across pages
+
 ## [2.0.0]
 
 ### DTS
@@ -39,7 +55,6 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Improve error reporting and stability
 - Update Zephyr OpenOCD and cross-platform build scripts; bump zinstaller to v2.0
 - Refine PowerShell policy, try to avoid powershell resctrictions on windows
->>>>>>> version-2
 
 ## [1.3.37]